Use isNaN for all number tests
This simplifies the code slightly, and it also allows humanise to do the
conversion to the number type it expects.

The primary reason for doing this is floats are being skipped because
they will never pass the test parseInt(input) !== input.  But as a side
effect strings that contain numbers will now also work.